+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Adds the `/default_format` command.
+- Adds the `/blocked` command.
 
 ## [v9.3.0](https://github.com/lexicalunit/spellbot/releases/tag/v9.3.0) - 2023-10-04